requirements.txt
picamera2>=0.3.9
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=10.0.0
python-telegram-bot>=20.3
pyyaml>=6.0
imageio>=2.31.0
RPi.GPIO>=0.7.1
adafruit-circuitpython-dht>=3.7.0
schedule>=1.2.0
python-dotenv>=1.0.0
astral>=3.2  # For sunrise/sunset calculations
